refactor: split BasePea in runtime and Pea
